Fix Stack configuration (#138)
* Fix Stack configuration

The `stack.yaml` configuration, using LTS 18.28 (GHC 8.10.7), specifies
use of a newer version of `hashable` than is provided by the snapshot.
This newer version is out of bounds of some other dependencies in the
snapshot.  Setting `allow-newer: true` is sufficient to resolve the
issue, however, as the dependencies work fine with the new version.

The `stack.yaml.lock` file was old, created using LTS 17.4 (GHC 8.10.4).
This commit includes the newer version that corresponds to building with
the current `stack.yaml`.

The `stack-aeson-2.yaml` configuration, using LTS 19.4 (GHC 9.0.2),
built fine.  This commit changes it to use LTS 19.33, the final snapshot
for that GHC version.

Note that `stack-aeson-2.yaml.lock` is not in the repository.  Perhaps
either it should be added or `stack.yaml.lock` should be removed, for
consistency.

* Remove stack.yaml.lock