Transaction class constraints (#147)
* Reset search_path before building database schema

This ensures fully qualified object names end up in the
representations, which is nice, but more importantly, it ensures
the schema is immune to the user's search_path setting.

* Add a migration that changes codd_admin's search_path

* Fix existing tests

* Adding test that ensures search_path is irrelevant

* InTxnT newtype and some hoisting functions added

Still very early work. The `InTxnT` ctor must be exposed for some
of the hoisting, and it's not clear how we can allow hoisting
monadic type parameters of custom types without doing so.

* Create `AppliedMigration` that does not have `(m :: * -> *)`

The monadic type parameter forced hoisting in lots of places where
it was purely unnecessary. This even helped clean up some of the
code.

* More adjustments, but still missing ResourceT+InTxnT harmony

A callback that runs in `InTxnT m` needs to be callable from more
than one base monad. In one branch of code, it's called from `m`,
but in another, it's called from `ResourceT m`. Its type is clearly
insufficient, so we'll probably need a class.

* Add `InTxn` class constraint and force universally quantified last
actions in the code.

This is a "cheap "escape from more restrictive monadic composition,
since now last actions must be able to run in _any_ monad that
satisfies a few class constraints, rather than constraints based
on the base monad.

This is fine for our codebase as it is, but is over-restrictive.

* Add `InTxn` constraint to function that actually requires it

* Make tests compile and pass

* Two parameter class approach, but still WIP

* lib stanza compiles, but lots of nasty trick to get there

- Lots of type applications to guide type inference
- Lots of apparently redundant class constraints
- Lots of `txn ~ InTxnT (ResourceT m)`, which should be unnecessary
- Pushing `MonadResource m` constraint up, beyond where needed

It's time to start tidying up

* Remove redundant MonadIO constraints

* Simplify a few more constraints, type arguments and hoisting

* Less explicit type applications, simpler functions, better docs

* Add more `InTxn` and `NotInTxn` instances

* No more need to export `InTxnT` constructor

* Enable Wredundant-constraints temporarily and simplify those

* Move MonadResource constraints around a bit

In general, this pushes MonadResource deeper into where it's
necessary, but inside the Internal module it is actually added
to one function where it isn't strictly necessary, but where it
better aligns monadic types with the other functions, reducing
the need for lifting

* Fixing `run format-hs` and fixing constraints until tests compile

* Remove `beginCommitTxnBracket` as it breaks the module's sandbox

It is possible to declare a function that is not constrained with
`NotInTxn m`, yet it called `beginCommitTxnBracket` by using
`liftIO`.

* Rename `CheckTxnFancy` to `CheckTxnWit`

* Relax `MonadUnliftIO` constraint to `MonadIO` in a few places

* resourcet no longer necessary for executable

* Beginning type-level sandbox checks

* Format haskell files

* Bring automatic rolling back on exception back for now

Otherwise retry tests fail, and that's something I prefer to fix
some other time.

* Ignore one more brittany error when formatting, format files

* Remove default-tx-read-only migration

With the new transaction class constraints, we can be much more
certain every transaction is read-write