Signicantly refactor block parsing
The previous code's separation between "can be first char",
"can be last char", and "inner contents parsing" was too confusing.

We keep "can be first char" as a small optimisation so we can still
use `takeWhile` instead of the backtracking that would ensue with
`many`, but otherwise resort to simple `Alternative` parsing with
backtracking where a block might be, and separate functions for
each kind of block.

This also implements parentheses parsing properly, and contains
a few other fixes.