Close file handles when file Streams are fully read
This still doesn't solve the issue, but will likely play an important
role, as code elsewhere won't have to worry about explicitly closing
files when they're done with them.

This shouldn't be a problem because we can't rewind file streams and
because `hClose` can be called more than once on the same handle.