Finally make at most one migration file be open at a time
This is not a pretty implementation.
Env vars to be templated are stored in each `SqlMigration` object (both
their names and values), and one of the functions that return migrations
and streams from disk higher up in the internal layers of the codebase
now reads that, closes the stream it had before and opens a new one.

That new FileStream is returned by calling the parser again and overriding
specific parts of the returned record, assuming the file won't change
in between the two moments it is opened (I don't think we can escape
this assumption regardless of implementation if we want to parse+report
errors in migrations before we start applying them).