Also register if a no-txn migration failed
This is necessary to allow resuming application of failed no-txn migrations
from their point of failure onwards.

While this should be extremely rare, if we are in the business of retrying
`BEGIN ... COMMIT` blocks of no-txn migrations, we're already in
extremely rare territory.

It is also behaviour that might be useful in the darkest of hours.

We will still need to change how we collect pending migrations, because now
a migration's applied state is no longer binary (i.e. applied or not),
implement seeking and a bunch of other things.