Changing code so we can insert into codd_schema even on sql statement errors
Right now this amounts to throwing exceptions in different places, and using
a new custom exception type to carry information of where the last `BEGIN`
was higher up in the call stack (this does not feel great), if there was one at all.
But so far we don't transmit that yet.

Also, this removes calls to retry statements in no-txn migrations
from functions deeper in the call stack, so that retrying is done higher up where
it's possible to re-read files and seek the stream to the right statement. And this
actually helped clean up some of the code.

Still much to do, though.