Return error types instead of using exceptions, make new test pass
This is a super specific error case of a failed bootstrapping no-txn migration
that runs enough statements to make the default connection string accessible.

With this commit codd will create codd_schema after this failure and
register this migration was partially applied.

This is the most unlikely and useless error case I can think of, and
the implementation messes with retry timeouts, so it's not clear this
is a good idea.

Plus, `codd add` might run into it. Since we have to address other issues
with `codd add` given the new code anyway, it might not be that big of a problem.

At least we can say codd will be helpful even in the most extreme cases.
Even if I can't think of a case where this particular feature will do that.